net: phy: do not soft reset PHY by default
For most PHY's a soft reset will revert them to power-on or strapped
defaults and un-do PHY configuration which was done in a prior boot
stage.

The Linux kernel at one time performed a PHY soft reset by default
and it was later removed and optionally done in PHY drivers for those
PHY's that needed it.

Follow this guideline by removing the call to phy_reset and where it is
needed a PHY driver can call it.

Signed-off-by: Tim Harvey <tharvey@gateworks.com>